Added a custom dictionary class to store the config dictionary of colorview2d.CvFig class.
The class is checking in __setitem__ whenever specific options are
changed and calls the appropriate plotting routines if there is an
active plot.
This way calls like

cvfig.config['Cbmin'] = 0.2
cvfig.config = {'Font': 'Ubuntu', 'Fontsize': 18}

work like a charm.

Another method, update_raw(), can be used to just update the dictionary.
This is meant to be invoked by init calls and the like, whenever we do
not want to update any existing plot.